+++ src/test/java/org/jabref/gui/externalfiles/AutoSetFileLinksUtilTest.java
@@ -49,7 +49,12 @@ public class AutoSetFileLinksUtilTest {
         List<LinkedFile> actual = util.findAssociatedNotLinkedFiles(entry);
         assertEquals(expected, actual);
     }
-
+    
+    /*
+    @Test
+    public void linkAssociatedFilesErrorDialog() throws Exception {
+        
+     */
     @Test
     public void testFindAssociatedNotLinkedFilesForEmptySearchDir() throws Exception {
         when(fileDirPrefs.isBibLocationAsPrimary()).thenReturn(false);