+++ CHANGELOG.md
@@ -14,6 +14,8 @@ Note that this project **does not** adhere to [Semantic Versioning](http://semve
 
 - We added a query parser and mapping layer to enable conversion of queries formulated in simplified lucene syntax by the user into api queries. [#6799](https://github.com/JabRef/jabref/pull/6799)
 - We added some basic functionality to customise the look of JabRef by importing a css theme file. [#5790](https://github.com/JabRef/jabref/issues/5790)
+- We added an error dialog, in case there are problems with linking filed [#5890]
+(https://github.com/JabRef/jabref/issues/5890)
 
 ### Changed